% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_random_walk.R
\name{bounded_walk}
\alias{bounded_walk}
\title{Create a bounded random walk}
\usage{
bounded_walk(n.times = 100, maxDist = 10, start = c(0, 0))
}
\arguments{
\item{n.times}{integer, the number of steps to take (default 100)}

\item{maxDist}{numeric, the maximum distance from starting posistion (default NA)}

\item{start}{x and y starting position (default (0,0))}
}
\description{
This function creates a random or semi-random walk bounded by a maximum distance from the starting position.
}
\examples{
bounded_walk()
}
\keyword{bounded}
\keyword{random}
\keyword{walk,}
