% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2mat.R
\name{dat2mat}
\alias{dat2mat}
\title{Convert a dataframe in a long format to a wide matrix}
\usage{
dat2mat(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
This function converts 3-column dataframe such as produced with mat2dat back into a matrix with the IDs in the first two columns used as dimnames
}
\examples{
mat2dat()
}
\keyword{rematrix}
