% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_kinship.R
\name{expected_kinship}
\alias{expected_kinship}
\title{Expected Kinship}
\usage{
expected_kinship(id = "id", mother_id = "mother_id",
  father_id = "father_id", sex = "sex", data)
}
\arguments{
\item{id}{character string specifying column of individuals ids}

\item{mother_id}{character string specifying column of mother ids}

\item{father_id}{character string specifying column of father ids}

\item{sex}{character string specifying column of individuals sexes}

\item{data}{dataframe}
}
\description{
This function calculates the expected kinship value based on maternal, paternal, and combined pedigrees
}
\examples{
mat2dat()
}
\keyword{kinship}
