% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_random_points.R
\name{fast_random_points}
\alias{fast_random_points}
\title{Sample relocations from a utilization distribution}
\usage{
fast_random_points(probweights, nd, dc, dgdf, gridrad)
}
\arguments{
\item{probweights}{a probability vector from which to select individuals, length must be greater than or equal to nd}

\item{nd}{number of individuals to sample}

\item{dc}{a matrix of coordinates}

\item{dgdf}{a dataframe of probabilities per individual in each cell}

\item{gridrad}{integer, half of the (square) gridsize}
}
\description{
This function samples relocations given a set of coordinates, number of individuals, and probability of each individual being relocated at each coordinate
}
\examples{

}
\keyword{relocations}
\keyword{sample}
