% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_assign.R
\name{group_assign}
\alias{group_assign}
\title{Assign individuals to groups based on 2-dimensional location and group size}
\usage{
group_assign(data = data, id = "id", xcoord = "x", ycoord = "y",
  time = "day", group_size = 2, method = "hclust")
}
\arguments{
\item{data}{list of simulation data generated by swap or UD method}

\item{id}{individual identifier field}

\item{time}{time step variable}

\item{group_size}{average group size in results}

\item{method}{grouping method, currently only hierarchical clustering implemented}

\item{xcoords}{x coordinates}

\item{ycoords}{y coordinates}
}
\description{
This function clusters individuals into groups based on their locations and some mean group size.
}
\examples{
group_assign()
}
\keyword{cluster}
\keyword{group}
\keyword{ids}
\keyword{ids,}
