% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_weight.R
\name{half_weight}
\alias{half_weight}
\title{Half Weight Index}
\usage{
half_weight(
  sightings = sightings,
  group_variable = group_variable,
  dates = dates,
  IDs = IDs,
  diag = FALSE,
  symmetric = TRUE,
  mask = NULL,
  assocInd = "HWI",
  M = 0.5
)
}
\arguments{
\item{sightings}{a dataframe in linear format, i.e., one row per individual per observation}

\item{group_variable}{a unique group identifier}

\item{dates}{dates}

\item{IDs}{individual IDs}

\item{diag}{logical, should the diagonal be included (default FALSE)}

\item{symmetric}{logical, should a symmetric matrix be returned (default TRUE)}

\item{mask}{an availability matrix where rows are individuals and columns are sighting dates, cells should be 1 for available or NA if not}

\item{assocInd}{the association index "HWI"}

\item{M}{the M-weight for the index, set to 0.5 for half-weight index}
}
\description{
This function calculates the half weight index for data in linear format
}
\examples{
dat2mat()
}
\keyword{association}
\keyword{half}
\keyword{index}
\keyword{weight}
