% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_perm.R
\name{node_perm}
\alias{node_perm}
\title{Node permutation test}
\usage{
node_perm(mod, nperm = 1000, offset = NA, plot = TRUE, export = FALSE)
}
\arguments{
\item{mod}{generalized linear model}

\item{nperm}{number of shuffles}

\item{offset}{names of offset variable to shuffle with response if present}

\item{plot}{default TRUE, to plot the distribution of random betas}
}
\description{
This function creates a random or semi-random walk bounded by a maximum distance from the starting position.
}
\examples{
node_perm()
}
\keyword{bounded}
\keyword{random}
\keyword{walk,}
