% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_pairs.R
\name{reduce_pairs}
\alias{reduce_pairs}
\title{A Function to Reduce Redundant Dyads}
\usage{
reduce_pairs(x, ID1, ID2)
}
\arguments{
\item{x}{a matrix or dataframe}

\item{ID1, ID2}{specifications of the columns containing the dyad identifiers}
}
\description{
This function allows you to merge data based pairs of IDs that can be in either order
}
\examples{
mhead()
merge_pairs()
}
\keyword{pairs}
\keyword{reduce}
