% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ratio.R
\name{simple_ratio}
\alias{simple_ratio}
\title{Simple Ratio Index}
\usage{
simple_ratio(
  sightings = sightings,
  group_variable = group_variable,
  dates = dates,
  IDs = IDs,
  diag = FALSE,
  symmetric = TRUE,
  mask = NULL,
  assocInd = "SRI"
)
}
\arguments{
\item{sightings}{a dataframe in linear format, i.e., one row per individual per observation}

\item{group_variable}{a unique group identifier}

\item{dates}{dates}

\item{IDs}{individual IDs}

\item{diag}{logical, should the diagonal be included (default FALSE)}

\item{symmetric}{logical, should a symmetric matrix be returned (default TRUE)}

\item{mask}{an availability matrix where rows are individuals and columns are sighting dates, cells should be 1 for available or NA if not}

\item{assocInd}{whether the association index "SRI", the number of sightings together "X", or total "XY" should be returned}
}
\description{
This function calculates the simple ratio index for data in linear format
}
\examples{
dat2mat()
}
\keyword{association}
\keyword{index}
\keyword{ratio}
\keyword{simple}
