% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test Function}
\usage{
test_function()
}
\description{
This function prepares students for the exercise "Using genome-wide SNP data to investigate social structure in pilot whales"
}
