% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2mat.R
\name{dat2mat}
\alias{dat2mat}
\title{Convert a dataframe in a long format to a wide matrix}
\usage{
dat2mat(data = data, forceSymmetric = FALSE, diag = FALSE)
}
\arguments{
\item{data}{a matrix}

\item{forceSymmetric}{boolean, if matrix has a empty triangle, fill over the diagonal}

\item{diag}{boolean, should a diagonal be added if absent?}
}
\description{
This function converts 3-column dataframe such as produced with mat2dat back into a matrix with the IDs in the first two columns used as dimnames
}
\details{
If value column is type numeric or integer the function will return a numeric matrix, otherwise a character matrix will be returned.
}
\examples{
mat2dat()
}
\keyword{rematrix}
