% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_assign.R
\name{group_assign}
\alias{group_assign}
\title{Assign individuals to groups based on 2-dimensional location and group size}
\usage{
group_assign(
  data = data,
  id = "id",
  xcoord = "x",
  ycoord = "y",
  time = NULL,
  group_size = NULL,
  group_vector = NULL,
  method = c("hclust", "kmeans")
)
}
\arguments{
\item{data}{list of simulation data generated by swap or UD method}

\item{id}{individual identifier field}

\item{time}{time step variable}

\item{group_size}{average group size in results}

\item{group_vector}{vector of group sizes per time step}

\item{method}{grouping method, either kmeans or hierarchical clustering}

\item{xcoords}{x coordinates}

\item{ycoords}{y coordinates}
}
\description{
This function clusters individuals into groups based on their locations and some mean group size.
}
\examples{
group_assign()
}
\keyword{cluster}
\keyword{group}
\keyword{ids}
\keyword{ids,}
